Move mouse stuff to interaction plugin
